Finalize adding recipes to site
The muffins page now contains the recipes with
ingredients and instructions broken out by part of
recipe. Also added a link to the source of the
recipe as I cannot take credit for it.

Removed the ramen recipe due to its irrelevance
to the site and incomplete recipe data.

Fixed a typo in one of the image files.